+++ classes/Gateways/check/Gateway.class.php
@@ -33,10 +33,6 @@ class Gateway extends \Shop\Gateway
      * @var string */
     protected $gw_provider = 'Pay by Check';
 
-    /** Flag this gateway as bundled with the Shop plugin.
-     * @var integer */
-    protected $bundled = 1;
-
 
     /**
      * Constructor.