+++ classes/Gateways/ppcheckout/Gateway.class.php
@@ -45,11 +45,6 @@ class Gateway extends \Shop\Gateway
      * @var string */
     protected $gw_desc = 'Checkout with PayPal';
 
-    /** Flag this gateway as bundled with the Shop plugin.
-     * Gateway version will be set to the Shop plugin's version.
-     * @var integer */
-    protected $bundled = 1;
-
     /** Paypal API URL, sandbox or production.
      * @var string */
     private $api_url;